graphutil header
